+++ src/GafferScene/EditScopeAlgo.cpp
@@ -39,6 +39,7 @@
 #include "GafferScene/Prune.h"
 #include "GafferScene/PathFilter.h"
 #include "GafferScene/SceneProcessor.h"
+#include "GafferScene/ShaderTweaks.h"
 #include "GafferScene/Transform.h"
 
 #include "Gaffer/EditScope.h"
@@ -48,8 +49,15 @@
 #include "Gaffer/StringPlug.h"
 
 #include "IECore/AngleConversion.h"
+#include "IECore/CamelCase.h"
 
 #include "OpenEXR/ImathMatrixAlgo.h"
+
+#include "boost/algorithm/string/replace.hpp"
+#include "boost/container/flat_map.hpp"
+#include "boost/tokenizer.hpp"
+
+#include <unordered_map>
 
 using namespace std;
 using namespace Imath;
@@ -264,4 +272,244 @@ boost::optional<GafferScene::EditScopeAlgo::TransformEdit> GafferScene::EditScop
 		row->cellsPlug()->getChild<Spreadsheet::CellPlug>( g_scale )->valuePlug<V3fPlug>(),
 		row->cellsPlug()->getChild<Spreadsheet::CellPlug>( g_pivot )->valuePlug<V3fPlug>()
 	};
+}
+
+// Shaders
+// =======
+
+namespace
+{
+
+SceneProcessorPtr shaderParameterProcessor( const std::string &attribute, const std::string &name )
+{
+	SceneProcessorPtr result = new SceneProcessor( name );
+
+	SpreadsheetPtr spreadsheet = new Spreadsheet;
+	result->addChild( spreadsheet );
+	spreadsheet->selectorPlug()->setValue( "${scene:path}" );
+
+	PathFilterPtr pathFilter = new PathFilter;
+	result->addChild( pathFilter );
+	pathFilter->pathsPlug()->setInput( spreadsheet->activeRowNamesPlug() );
+
+	ShaderTweaksPtr shaderTweaks = new ShaderTweaks;
+	result->addChild( shaderTweaks );
+	shaderTweaks->inPlug()->setInput( result->inPlug() );
+	shaderTweaks->filterPlug()->setInput( pathFilter->outPlug() );
+	shaderTweaks->enabledPlug()->setInput( result->enabledPlug() );
+	shaderTweaks->shaderPlug()->setValue( attribute );
+	shaderTweaks->localisePlug()->setValue( true );
+	shaderTweaks->ignoreMissingPlug()->setValue( true );
+
+	auto rowsPlug = static_cast<Spreadsheet::RowsPlug *>( PlugAlgo::promoteWithName( spreadsheet->rowsPlug(), "edits" ) );
+	Metadata::registerValue( rowsPlug, "spreadsheet:defaultRowVisible", new BoolData( false ) );
+	Metadata::registerValue( rowsPlug->defaultRow(), "spreadsheet:rowNameWidth", new IntData( 300 ) );
+
+	result->outPlug()->setInput( shaderTweaks->outPlug() );
+
+	return result;
+}
+
+/// \todo Create a central renderer/attribute registry that we can
+/// query for this information.
+const boost::container::flat_map<string, string> g_rendererAttributePrefixes = {
+	{ "ai", "Arnold" },
+	{ "dl", "Delight" },
+	{ "as", "Appleseed" },
+	{ "gl", "OpenGL" },
+	{ "osl", "OSL" },
+	{ "ccl", "Cycles" }
+};
+
+string parameterProcessorName( const std::string &attribute )
+{
+	string rendererPrefix;
+	vector<string> parts;
+
+	using Tokenizer = boost::tokenizer<boost::char_separator<char>>;
+	Tokenizer tokens( attribute, boost::char_separator<char>( ":" ) );
+	for( const auto &token : tokens )
+	{
+		if( parts.empty() )
+		{
+			auto it = g_rendererAttributePrefixes.find( token );
+			if( it != g_rendererAttributePrefixes.end() )
+			{
+				rendererPrefix = it->second;
+				continue;
+			}
+		}
+		CamelCase::split( token, back_inserter( parts ) );
+	}
+
+	return rendererPrefix + CamelCase::join( parts.begin(), parts.end() ) + "Edits";
+}
+
+SceneProcessor *acquireParameterProcessor( EditScope *editScope, const std::string &attribute, bool createIfNecessary )
+{
+	static unordered_map<string, string> attributeProcessors;
+	auto inserted = attributeProcessors.insert( { attribute, "" } );
+	if( inserted.second )
+	{
+		const string name = parameterProcessorName( attribute );
+		EditScope::registerProcessor(
+			name,
+			[ attribute, name ] () {
+				return ::shaderParameterProcessor( attribute, name );
+			}
+		);
+		inserted.first->second = name;
+	}
+
+	return editScope->acquireProcessor<SceneProcessor>( inserted.first->second, createIfNecessary );
+}
+
+ConstDataPtr parameterValue( const ScenePlug *scene, const ScenePlug::ScenePath &path, const std::string &attribute, const IECoreScene::ShaderNetwork::Parameter &parameter )
+{
+	if( !scene->exists( path ) )
+	{
+		string pathString; ScenePlug::pathToString( path, pathString );
+		throw IECore::Exception( boost::str( boost::format( "Location \"%1%\" does not exist" ) % pathString ) );
+	}
+
+	auto attributes = scene->fullAttributes( path );
+	auto shaderNetwork = attributes->member<IECoreScene::ShaderNetwork>( attribute );
+	if( !shaderNetwork )
+	{
+		throw IECore::Exception( boost::str( boost::format( "Attribute \"%1%\" does not exist" ) % attribute ) );
+	}
+
+	const IECoreScene::Shader *shader;
+	if( parameter.shader.string().size() )
+	{
+		shader = shaderNetwork->getShader( parameter.shader );
+		if( !shader )
+		{
+			throw IECore::Exception( boost::str( boost::format( "Shader \"%1%\" does not exist" ) % parameter.shader ) );
+		}
+	}
+	else
+	{
+		shader = shaderNetwork->outputShader();
+		if( !shader )
+		{
+			throw IECore::Exception( "Output shader does not exist" );
+		}
+	}
+
+	const Data *result = shader->parametersData()->member( parameter.name );
+	if( result )
+	{
+		return result;
+	}
+	else
+	{
+		throw IECore::Exception( boost::str( boost::format( "Parameter \"%1%\" does not exist" ) % parameter.name ) );
+	}
+}
+
+} // namespace
+
+bool GafferScene::EditScopeAlgo::hasParameterEdit( const Gaffer::EditScope *scope, const ScenePlug::ScenePath &path, const std::string &attribute, const IECoreScene::ShaderNetwork::Parameter &parameter )
+{
+	return acquireParameterEdit( const_cast<EditScope *>( scope ), path, attribute, parameter, /* createIfNecessary = */ false );
+}
+
+TweakPlug *GafferScene::EditScopeAlgo::acquireParameterEdit( Gaffer::EditScope *scope, const ScenePlug::ScenePath &path, const std::string &attribute, const IECoreScene::ShaderNetwork::Parameter &parameter, bool createIfNecessary )
+{
+	string pathString;
+	ScenePlug::pathToString( path, pathString );
+
+	// If we need to create an edit, we'll need to do a compute to figure our the parameter
+	// type and value. But we don't want to do that if we already have an edit. And since the
+	// compute could error, we need to get the parameter value before making _any_ changes, so we
+	// don't leave things in a partial state. We use `ensureParameterValue()` to get the value
+	// lazily at the first point we know it will be needed.
+	ConstDataPtr parameterValue;
+	auto ensureParameterValue = [&] {
+		if( !parameterValue )
+		{
+			parameterValue = ::parameterValue( scope->outPlug<ScenePlug>(), path, attribute, parameter );
+		}
+	};
+
+	// Find processor, and row for `path`.
+
+	auto *processor = acquireParameterProcessor( scope, attribute, /* createIfNecessary = */ false );
+	if( !processor )
+	{
+		if( !createIfNecessary )
+		{
+			return nullptr;
+		}
+		else
+		{
+			ensureParameterValue();
+			processor = acquireParameterProcessor( scope, attribute, /* createIfNecessary = */ true );
+		}
+	}
+
+	auto *rows = processor->getChild<Spreadsheet::RowsPlug>( "edits" );
+	Spreadsheet::RowPlug *row = rows->row( pathString );
+	if( !row )
+	{
+		if( !createIfNecessary )
+		{
+			return nullptr;
+		}
+		ensureParameterValue();
+		row = rows->addRow();
+		row->namePlug()->setValue( pathString );
+	}
+
+	// Find cell for parameter
+
+	string tweakName = parameter.name.string();
+	if( parameter.shader.string().size() )
+	{
+		tweakName = parameter.shader.string() + "." + tweakName;
+	}
+
+	string columnName = boost::replace_all_copy( tweakName, ".", "_" );
+	if( auto *cell = row->cellsPlug()->getChild<Spreadsheet::CellPlug>( columnName ) )
+	{
+		return cell->valuePlug<TweakPlug>();
+	}
+
+	if( !createIfNecessary )
+	{
+		return nullptr;
+	}
+
+	// No tweak for parameter yet. Create it.
+
+	ensureParameterValue();
+	ValuePlugPtr valuePlug = PlugAlgo::createPlugFromData( "value", Plug::In, Plug::Default, parameterValue.get() );
+	/// \todo The TweakPlug constructor makes a plug where `namePlug()` has a default value of "" and `enabledPlug()`
+	/// has a default value of `true`. This makes for a lot of redundancy in our spreadsheet serialisations because
+	/// every cell will have a `setValue()` for the name, and we expect to have fewer enabled cells than disabled ones.
+	/// Change the TweakPlug constructor (or provide an overload) so we can get the defaults we want. Consider the
+	/// relationship to NameValuePlug and ShufflePlug constructors at the same time.
+	TweakPlugPtr tweakPlug = new TweakPlug( tweakName, valuePlug, TweakPlug::Replace, false );
+
+	auto *shaderTweaks = processor->getChild<ShaderTweaks>( "ShaderTweaks" );
+	shaderTweaks->tweaksPlug()->addChild( tweakPlug );
+
+	size_t columnIndex = rows->addColumn( tweakPlug.get(), columnName, /* adoptEnabledPlug = */ true );
+	tweakPlug->setInput( processor->getChild<Spreadsheet>( "Spreadsheet" )->outPlug()->getChild<Plug>( columnIndex ) );
+
+	return row->cellsPlug()->getChild<Spreadsheet::CellPlug>( columnIndex )->valuePlug<TweakPlug>();
+}
+
+void GafferScene::EditScopeAlgo::removeParameterEdit( Gaffer::EditScope *scope, const ScenePlug::ScenePath &path, const std::string &attribute, const IECoreScene::ShaderNetwork::Parameter &parameter )
+{
+	TweakPlug *edit = acquireParameterEdit( scope, path, attribute, parameter, /* createIfNecessary = */ false );
+	if( !edit )
+	{
+		return;
+	}
+	// We're unlikely to be able to delete the row or column,
+	// because that would affect other edits, so we simply disable
+	// the edit instead.
+	edit->enabledPlug()->setValue( false );
 }